Refactor the validators code
