+++ middleware/validatorMiddleware.js
@@ -1,0 +1,14 @@
+'use strict';
+const { validationResult } = require('express-validator');
+
+// @desc Finds the validation errors in this request and wraps them in an object with handy functions
+
+const validatorMiddleware = (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+  next();
+};
+
+module.exports = validatorMiddleware;
