+++ routes/categoryRoutes.js
@@ -2,7 +2,6 @@
 
 const express = require('express');
 const router = express.Router();
-const { param, validationResult } = require('express-validator');
 
 const {
   getCategories,
@@ -12,6 +11,12 @@ const {
   deleteCategory,
 } = require('../services/categoryService');
 
+const {
+  getCategoryValidator,
+  updateCategoryValidator,
+  deleteCategoryValidator,
+} = require('../utils/validators/categoryValidators');
+
 // Explain the meaning of middleware : https://expressjs.com/en/guide/writing-middleware.html
 // router.get('/', getCategories);
 // router.post('/', createCategory);
@@ -20,20 +25,8 @@ router.route('/').get(getCategories).post(createCategory);
 
 router
   .route('/:id')
-  .get(
-    // 1- rules
-    param('id').isMongoId().withMessage('Invalid category id'),
-    // 2- middleware => catch errors form rules if exist
-    (req, res) => {
-      // Finds the validation errors in this request and wraps them in an object with handy functions
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
-      }
-    },
-    getCategory
-  )
-  .put(updateCategory)
-  .delete(deleteCategory);
+  .get(getCategoryValidator, getCategory)
+  .put(updateCategoryValidator, updateCategory)
+  .delete(deleteCategoryValidator, deleteCategory);
 
 module.exports = router;
