+++ server.js
@@ -6,7 +6,7 @@ const ApiError = require('./utils/ApiErorr');
 //=====Routes
 const categoryRoutes = require('./routes/categoryRoutes');
 const dbConnection = require('./config/database');
-const globalErorrHandlingMidleware = require('./middleware/errorMidleware');
+const globalErorrHandlingMidleware = require('./middleware/errorMiddleware');
 dotenv.config({ path: 'config.env' });
 
 // express app
